requirements.txt
weaviate-client[agents]>=4.8.1
python-dotenv>=1.0.1
streamlit>=1.35.0
